+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ## 1.8 (not released yet)
+- introduce CASC_VAULT_PATHS to allow multiple vault paths to read from. CASC_VAULT_PATH kept for backwards compatibility and offering multi path too now.
 
 ## 1.7